fork: initialize EnduroSat fork of libcsp
Add EnduroSat attribution, virtual topology subsystem, Windows arch
port, traceroute and RF interface, CI infrastructure, and restructured
drivers layout. Remove upstream CI workflows and legacy samples.

+++ include/csp/interfaces/csp_if_can.h
@@ -162,9 +162,9 @@ extern "C" {
 /**
  * Fields used to uniquely define a CSP packet within each fragment header
  */
-#define CFP2_ID_CONN_MASK ((CFP2_DST_MASK << CFP2_DST_OFFSET) | \
-						   (CFP2_SENDER_MASK << CFP2_SENDER_OFFSET) | \
-					 (CFP2_PRIO_MASK << CFP2_PRIO_OFFSET) | \
+#define CFP2_ID_CONN_MASK (((uint32_t)CFP2_DST_MASK << CFP2_DST_OFFSET) | \
+						   ((uint32_t)CFP2_SENDER_MASK << CFP2_SENDER_OFFSET) | \
+					       ((uint32_t)CFP2_PRIO_MASK << CFP2_PRIO_OFFSET) | \
 						   (CFP2_SC_MASK << CFP2_SC_OFFSET))
